Sitnice

+++ WebProjekatPoslednjaVerzija/src/main/java/com/projekat/WebProjekat/service/ArtikalService.java
@@ -3,6 +3,7 @@ package main.java.com.projekat.WebProjekat.service;
 import main.java.com.projekat.WebProjekat.dto.ArtikalDto.ArtikalDto;
 import main.java.com.projekat.WebProjekat.entity.Artikal;
 import main.java.com.projekat.WebProjekat.entity.Menadzer;
+import main.java.com.projekat.WebProjekat.entity.Restoran;
 import main.java.com.projekat.WebProjekat.repository.ArtikalRepository;
 import main.java.com.projekat.WebProjekat.repository.RestoranRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,6 +32,9 @@ public class ArtikalService {
 
     public List<Artikal> findAll() { return artikalRepository.findAll(); }
 
+    public Restoran saveRestoran(Restoran restoran){
+        return restoranRepository.save(restoran);
+    }
     public Artikal dodajArtikal(ArtikalDto dto, Menadzer menadzer, String fileName){
         Artikal artikal = new Artikal(dto.getNaziv(), dto.getCena(), dto.getTip(), dto.getKolicina(), dto.getOpis(), menadzer.getRestoran());
         artikal.setPhotos(fileName);
@@ -44,7 +48,34 @@ public class ArtikalService {
         return artikal;
     }
 
-    public void delete(Artikal artikal) {
-        artikalRepository.delete(artikal);
+    public Artikal update(Long id, ArtikalDto artikalDto){
+        Artikal artikal = this.findOne(id);
+        if(!artikalDto.getNaziv().isEmpty()){
+            artikal.setNaziv(artikalDto.getNaziv());
+        }
+        if(artikalDto.getCena() > 0){
+            artikal.setCena(artikalDto.getCena());
+        }
+        if(artikalDto.getTip() != null){
+            artikal.setTip(artikalDto.getTip());
+        }
+        if(artikalDto.getKolicina() > 0){
+            artikal.setKolicina(artikalDto.getKolicina());
+        }
+        if(!artikalDto.getOpis().isEmpty()){
+            artikal.setOpis(artikalDto.getOpis());
+        }
+        return this.save(artikal);
+    }
+
+    public void delete(Long id, Restoran restoran) {
+        for (Artikal artikal : restoran.getArtikliUPonudi()) {
+            if (artikal.getId().equals(id)) {
+                restoran.getArtikliUPonudi().remove(artikal);
+                artikal.setRestoran(null);
+                artikalRepository.delete(artikal);
+                this.saveRestoran(restoran);
+            }
+        }
     }
 }
